Final version.

+++ src/core/texture/texture.h
@@ -16,12 +16,12 @@ public:
 	std::string name;
 	uint16_t width;
 	uint16_t height;
-	uint8_t wrap_s;
-	uint8_t wrap_t;
-	uint8_t filter_min;
-	uint8_t filter_max;
-	uint8_t internal_format;
-	uint8_t image_format;
+	uint16_t wrap_s;
+	uint16_t wrap_t;
+	uint16_t filter_min;
+	uint16_t filter_max;
+	uint16_t internal_format;
+	uint16_t image_format;
 };
 
 #endif
